+++ src/StorybookInit.tsx
@@ -5,13 +5,15 @@ import { InputStory } from '@components/Input/InputStory';
 import { InputV2Story } from '@components/InputV2/InputV2Story';
 import { NotificationStory } from '@components/Notification/NotificationStory';
 import { PopupStory } from '@components/Popup/PopupStory';
+import { ScnStory } from '@components/Scn/ScnStory';
 import { SkeletonStory } from '@components/Skeleton/story/SkeletonStory';
 import { SpinnerStory } from '@components/Spinner/SpinnerStory';
 import { Storybook } from '@components/Storybook';
 import { SwitchScgScnStory } from '@components/SwitchScgScn/SwitchScgScnStory';
 import { TextareaStory } from '@components/Textarea/TextareaStory';
 import { TooltipStory } from '@components/Tooltip/TooltipStory';
-import { createGlobalStyle } from 'styled-components';
+import { defaultLight } from '@constants/theme';
+import { createGlobalStyle, ThemeProvider } from 'styled-components';
 
 const Global = createGlobalStyle`
   body {
@@ -21,7 +23,7 @@ const Global = createGlobalStyle`
 
 export const StorybookInit = () => {
   return (
-    <>
+    <ThemeProvider theme={defaultLight}>
       <Global />
       <Storybook>
         <ButtonStory />
@@ -36,7 +38,8 @@ export const StorybookInit = () => {
         <DropdownStory />
         <PopupStory />
         <NotificationStory />
+        <ScnStory />
       </Storybook>
-    </>
+    </ThemeProvider>
   );
 };