+++ src/components/InputSearch/index.ts
@@ -1,1 +1,1 @@
-export { default as InputSearch } from './InputSearch';
+export * from './InputSearch';