+++ src/components/InputV2/index.ts
@@ -1,2 +1,2 @@
-export { default as InputV2 } from './InputV2';
+export * from './InputV2';
 export type { InputV2Props } from './types';