+++ src/components/Scn/index.tsx
@@ -1,3 +1,2 @@
 export * from './Scn';
-export * from './model';
 export * from './constants';